New structures, working with GL.
First, I got rid of the class for agents. It was heavy, considering all I wanted to do was reference some values. Agents became structs (which are a wrapper around vectors, as far as I can tell), and their fields became a hash.

This meant that functions that operate on agents could be parameterized over the (current-agent), making it possible to more easily implement

(ask-turtles
  (right 1)
  (move 1))

as something that expands to

(loop ...
  (parameterize ([current-agent agent])
    (right (current-agent) 1)
    (move  (current-agent) 1)))

or thereabouts. This is much simpler than the dance I had to go through in order to get through the interface on a class representing the agents.

Agents are now stored in a vector instead of a hash table. I'm assuming the performance differences are minimal; it seems like at 5K agents, update is under 1/60th of a second; at 10,000 agents, they take longer, and therfore a 1/60th tick is actually a problem. Either way, this is good for getting on with; most simulations will be the grid and some number of agents that are probably smaller than 10K. (If that's my current limit for "snappy real-time interaction," I'm happy.

No doubt, if I thread this, I'll run into a crash with OpenGL. Drawing is spread across multiple functions with no protections.