Cleaner approach.
So, if you think about your data, it becomes easier...

Instead of trying to do crazy macro-fu, I've created a few datatypes.

This makes it much easier. There are agentsets, and agents. Now, the macros
operate on agentsets, and there are functions (eg. filtering functions) that
consume and return agentsets. I was doing this in an ad-hoc way before,
 but this makes much more sense.

(ask _agentset_ _bodies_ ...)

is the macro, instead of

(ask-agentset _bodies_ ...)

which led to a proliferation of macros.