Changed the agentset rep.
To simplify create-breed, I've wrapped agentsets in a thunk.

This lets

(create-breed fish fishes)

define

fishes

as a procedure that, when invoked, does

(hash-ref agentsets 'fish)

and thus returns the agentset struct. This involved reworking most things so that they correctly de-thunked agentsets. However, the macros still work, and now
I'm creating turtles and patches using the same tools that users will use. This means that turtles and patches are not special. (Patches may need to become special... I'm not yet sure.)

However, if I can keep turtles and patches as not-special, that means that there's a default breed called turtles, and a default breed called patches. That's better than them being magical. All the set/get/ask forms need to work on turtles as well as breeds, so this is the right design path.