Now using dirty-patches.
I don't know if this is an improvement... well, it is, but it adds complexity.

There are now 'patches and 'dirty-patches.

When a patch is 'set', the dirty? flag is set on that patch. And, at that point, the patch is moved into the 'dirty-patches agentset.

When a turtle does a get on a patch, first there is a check to see if it is in the irty-patches set. If it is, the turtle reads that value. Otherwise, it reads from the patches set.

When a patch is cleaned (set dirty? to false), I need to move the fields from the dirty patch to the clean patch set. I'm pretty sure that isn't being handled.

However, if lets me have large worlds and run quickly. At the point that a 200x200 world is using all of the patches (40,000 agents), it will run slowly... until I figure out how to optimize everything.