A functional but sloppy sniff.
It could probably be much better. I just added the plural to the agent fields, and hope that everything works everywhere... by adding it to the defaults, it should be fine.  I needed this to get from the agent to the agentset, which apparently I had not done before.

Sniffing seems to be working.