Moving things around
This still isn't pretty, by any stretch. However, I got rid of a file, and moved things into slightly more sensible places.

I liked the restructuring in the db-backing branch, but it was too slow. It says to me that rethinking the underlying datastructures, and then providing clean interfaces, would do me a world of good.