Thinking about structs.
I've now explored enough to have a better sense for what I want for this system.

There really should be a clean way to introduce the types for breeds, and still have a consistent universe underneath. In other words, getting the data types right matters in the microworld, both for clarity, and for performance. (Or, notional performance.)

But, really, I just would prefer that things were neater. Structs are essentially vectors, so they should be performant. And, I'd like the named interface to agent values.