Have a world again.
Now, sniff.

Sniff is the bottleneck. I need to be able to quickly find my neighbors. And, for that matter, do the filtering over at least 500-1000 agents. If I can get up to 5000 agents, I'd be really happy. But, for now, being able to run 1000 agents at a respectable (30-60fps) framerate would be lovely.