Reasonable performance at 1000 agents.
At 500 agents, I get a framerate (using sniff, on every agent) that looks to be around 30fps. It might be a touch slower, but I haven't implemented a framerate calculation in world.rkt.

Point is, the 500 agents are cruizing around fluidly.

At 1000 agents, it's a touch slower. Now, the question is whether there is a way to speed up the query on the quadtree. Or, if changing the number of agents at each level of the tree might help? I'll have to read up on performance tuning quadtrees.

This is a good step. Reducing the abstractions involved in the microworld made a difference. In truth, the macro expansion involved is simpler, and it will make writing the various netlogo interface functions easier as well. In fact, they're really just functions now, and have a much simpler implementation pathway... some of them were going to have to have macro elements previously, just to be able to interop with all the structs.

This looks reasonably good.