Wolves/Sheep work.
Now, the question is... should I make the changes to so agentsets (when users are writing code, etc.) are just lists of numbers?

This might simplify a few things. For now, the examples I have run, which means the ADT is (probably) complete.

I'm going to go back to doing boids correctly, and see how performance looks with 1000-5000 boids all doing cohesion/avoidance/etc. correctly.