Boids works well.
This implements Reynolds's flocking algorithm reasonably correctly.

There are things that might still be missing from where I was, but it is much better than it used to be. I can sniff with a reasonably large number of agents, add and remove, and now need to think about bringing the interface back in.

Logging/plotting could probably be more efficient.