Improving direction/movement for bounce mode
Choosing to represent direction as a heading as opposed to a vector quantity was perahps a poor choice.

The movement *should* now be correct in the general case, but there's a dual degree/vector representation.
This way, when implementing edge bounces, the calculation can be carried out on the vectors (thus
flipping the x or y component), and the degree heading can be updated accordingly. Similarly,
turns (eg. (right 30)) can be computed on the degree heading, and the vectors updated.

This is certainly duplicating information, and will no doubt bite me in the rear later.
However, it seemed like the best, next incremental step.

Next is to clean up the directory, so that things that don't build are in a branch (as
woudl be proper.)